Solution in swift
Wrote it in swift on the website because I was working with swift on a
project at the time I was solving the challenge. Unable to currently
commit as a swift file because I'm on a windows machine at the time of
pushing the commit, so I'll just push it as a text file for now.